Refactor styles and related products carousel
